==> 1a_Data-Prep.ipynb <==
# Terrorism Trends and Determinants

# Data Management/Investigation
import wbdata
import pandas as pd
import numpy as np
#import country_converter as coco
import missingno as miss
import plotly.express as px
import seaborn as sns
import matplotlib.pyplot as plt
import datetime

# GTD Dataset

gtd = pd.read_excel("Data/globalterrorismdb_0919dist.xlsx")

gtd

gtd1993 = pd.read_excel("Data/gtd1993_0919dist.xlsx")
gtd1993

# Check if the column names match
list(gtd.columns)==list(gtd1993.columns)

# Consolidate
gtdfinal = gtd.append(gtd1993)
gtdfinal

# export preprocessed file
gtdfinal.to_csv('Data/gtdfinal.csv')
