==> 5_ARCH GARCH Models.Rmd <==

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r loads, warning=FALSE, include = FALSE}

#Load libraries
library(dplyr)
library(lubridate)
library(stringr)
library(ggplot2)
library(tidyr)
library(knitr)

library(tidyverse)
library(ggplot2)
library(forecast)
library(astsa) 
library(xts)
library(tseries)
library(fpp2)
library(fma)
library(lubridate)
library(TSstudio)
library(quantmod)
library(tidyquant)
library(plotly)
```


```{r}
sp <- getSymbols("^GSPC",auto.assign = FALSE, from = "1970-01-01",to="2018-12-31",src="yahoo")
head(sp)
start(sp)
end(sp)

sp=data.frame(sp)
sp <- data.frame(sp,rownames(sp))
head(sp)

colnames(sp)[7] = "date"
head(sp)

sp$date<-as.Date(sp$date,"%Y-%m-%d")
str(sp)

sp %>%
  ggplot()+
  geom_line(aes(y=GSPC.Close,x=date))+
  labs(title="S&P 500 Stock Market Index (1970 to 2018)",x="Date",y="Closing Price") +
  theme_bw()

plot_ly(sp, type='scatter',mode='lines') %>%
  add_trace(x = ~date, y = ~GSPC.Close, name="S&P Closing Price")
```


```{r}
# we define return as difference in logged values
sp.returns=ts(diff(log(sp$GSPC.Close))*100)
# autoplot(sp.returns,color="blue")+
#   xlab("Date") + ylab("Return") +
#   ggtitle("Daily S&P 500 Returns (1970 to 2018)")

sp %>%
  mutate(GSPC.Close=c(NA,diff(log(sp$GSPC.Close))*100)) %>%
  ggplot()+
  geom_line(aes(y=GSPC.Close,x=date))+
  labs(title="Daily S&P 500 Returns (1970 to 2018)",x="Date",y="Return") +
  theme_bw()
```


```{r}
ts(sp.returns,start=c(1970,1)) %>% ggtsdisplay(main="Observed S&P 500 Returns")
ts(sp.returns) %>% ggtsdisplay(main="Observed S&P 500 Returns")
```


```{r}
## empty list to store model fits
ARMA_res <- list()
## set counter
cc <- 1
## loop over AR
for (p in 0:4) {
  ## loop over MA
  for (q in 0:2) {
    if (p+q<7) {
    ARMA_res[[cc]] <- Arima(sp.returns, order = c(p,0,q))
    cc <- cc + 1
    }
  }
}

```

```{r}
## get AIC values for model evaluation
ARMA_AIC <- sapply(ARMA_res, function(x) x$aic)
ARMA_res[[which(ARMA_AIC == min(ARMA_AIC))]]
```

```{r}
## get BIC values for model evaluation
ARMA_BIC <- sapply(ARMA_res, function(x) x$bic)
ARMA_res[[which(ARMA_BIC == min(ARMA_BIC))]]
```


```{r}
plot(rstandard(Arima(sp.returns,order=c(4,0,1))),
     main="ARIMA(4,0,1) Model of S&P 500 Returns (1970 to 2018)",
     ylab="Standardized Residuals",
     xlab="")
```

```{r}
returns.arima.fit = Arima(sp.returns,order=c(4,0,1))
summary(returns.arima.fit)

returns.arima.fit.res <- returns.arima.fit$residuals
```

```{r}
returns.arima.fit.res %>% ggtsdisplay(main="Residuals")
returns.arima.fit.res^2 %>% ggtsdisplay(main="Squared residuals")
```

```{r}
library(FinTS)
ArchTest(returns.arima.fit.res,lags=1,demean=TRUE)
```

```{r}
model <- list() ## set counter
cc <- 1
for (p in 0:2) {
  for (q in 0:2) {
    if (p+q != 0){
  
model[[cc]] <- garch(na.omit(returns.arima.fit.res),order=c(q,p),trace=F)
cc <- cc + 1
}
}
} 

## get AIC values for model evaluation
GARCH_AIC <- sapply(model, AIC) ## model with lowest AIC is the best
model[[which(GARCH_AIC == min(GARCH_AIC))]]
```

```{r}
summary(final.fit <- garchFit(~garch(1,2), na.omit(returns.arima.fit.res)))

```

```{r}
summary(final.fit <- garchFit(~arma(4,1)+garch(1,2), data=sp.returns, trace=FALSE))
```
`
```{r}
summary(final.fit <- garchFit(~arma(2,1)+garch(1,1), data=sp.returns, trace=FALSE))
```
